Moved NcssCountRule to codesize.xml
