+++ pmd-java/src/main/resources/rulesets/java/codesize.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 
 <ruleset name="Code Size"
-    xmlns="http://pmd.sourceforge.net/ruleset/2.0.0"
+    xmlns="http://pmd.sourceforge.net/ruleset/3.0.0"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0 http://pmd.sourceforge.net/ruleset_2_0_0.xsd">
+    xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/3.0.0 http://pmd.sourceforge.net/ruleset_3_0_0.xsd">
 
   <description>
 The Code Size ruleset contains rules that find problems related to code size or complexity.
@@ -365,6 +365,7 @@ public class Person {   // this is more manageable
 
 <rule name="NcssMethodCount"
       message="The method {0}() has an NCSS line count of {1}"
+      deprecated="true"
      since="3.9"
      class="net.sourceforge.pmd.lang.java.rule.codesize.NcssMethodCountRule"
      externalInfoUrl="${pmd.website.baseurl}/pmd_rules_java_codesize.html#ncssmethodcount">
@@ -394,6 +395,7 @@ public class Foo extends Bar {
 <rule name="NcssTypeCount"
       message="The type has an NCSS line count of {0}"
      since="3.9"
+      deprecated="true"
      class="net.sourceforge.pmd.lang.java.rule.codesize.NcssTypeCountRule"
      externalInfoUrl="${pmd.website.baseurl}/pmd_rules_java_codesize.html#ncsstypecount">
     <description>
@@ -422,6 +424,7 @@ public class Foo extends Bar {
 <rule name="NcssConstructorCount"
       message="The constructor with {0} parameters has an NCSS line count of {1}"
       since="3.9"
+      deprecated="true"
       class="net.sourceforge.pmd.lang.java.rule.codesize.NcssConstructorCountRule"
       externalInfoUrl="${pmd.website.baseurl}/pmd_rules_java_codesize.html#ncssconstructorcount">
     <description>
@@ -447,6 +450,50 @@ public class Foo extends Bar {
 ]]>
    </example>
 </rule>
+
+    <rule name="NcssCount"
+          message="The {0} ''{1}'' has a NCSS line count of {2}."
+          since="3.9"
+          class="net.sourceforge.pmd.lang.java.rule.codesize.NcssCountRule"
+          metrics="true"
+          externalInfoUrl="${pmd.website.baseurl}/pmd_rules_java_codesize.html#ncsscount">
+        <description>
+            This rule uses the NCSS (Non-Commenting Source Statements) metric to determine the number of lines
+            of code in a class, method or constructor. NCSS ignores comments, blank lines, and only counts actual
+            statements. For more details on the calculation, see the documentation of
+            the [NCSS metric](/pmd_java_metrics_index.html#non-commenting-source-statements-ncss).
+        </description>
+        <priority>3</priority>
+        <example>
+            <![CDATA[
+import java.util.Collections;       // +0
+import java.io.IOException;         // +0
+
+class Foo {                         // +1, total Ncss = 12
+
+  public void bigMethod()           // +1
+      throws IOException {
+    int x = 0, y = 2;               // +1
+    boolean a = false, b = true;    // +1
+
+    if (a || b) {                   // +1
+      try {                         // +1
+        do {                        // +1
+          x += 2;                   // +1
+        } while (x < 12);
+
+        System.exit(0);             // +1
+      } catch (IOException ioe) {   // +1
+        throw new PatheticFailException(ioe); // +1
+      }
+    } else {
+      assert false;                 // +1
+    }
+  }
+}
+]]>
+        </example>
+    </rule>
 
     <rule
         name="TooManyMethods"
