+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/metrics/MetricsRulesTest.java
@@ -26,7 +26,6 @@ public class MetricsRulesTest extends SimpleAggregatorTst {
     @Override
     public void setUp() {
         addRule(RULESET, "CyclomaticComplexity");
-        addRule(RULESET, "NcssCount");
         addRule(RULESET, "NPathComplexity");
         addRule(RULESET, "DataClass");
     }
